backend/requirements.txt
nostril
pyenchant
